[MERGE #1588 @Cellule] Merge WebAssembly to Master
Merge pull request #1588 from WebAssembly_Stage

This is the first step to bring WebAssembly to the product.
Our implementation is not currently complete, but working directly in master will alleviate some overhead for the people working on WebAssembly and help verify the code's stability through the normal testing pipeline.
These changes didn't cause any noticeable perf regressions when testing in IB and on my machine.