[MERGE #1784 @ThomsonTan] Disable tag checks optimization for BailOutOnTaggedValue
Merge pull request #1784 from ThomsonTan:FixOptTagChecks

For BailOutOnTaggedValue, we could not have a chance to update profile when we get tagged value because profile will only be produced when defining the value, not using them. So even rejit after some bailouts will not help because of missing profile on the value type.  This change disable tag checks for rejit when BailOutOnTaggedValue is detected because disable it doesn't hurt perf when the value type is not fixed.