[MERGE #1897 @agarwal-sandeep] Optimize deleting last property on an object
Merge pull request #1897 from agarwal-sandeep:optimizedeletelast

Today if we are in PathTypeHandler and we delete a property we go to
SimpleDictionaryTypeHandler. Having a type in SimpleDictionaryTypeHandler
is expensive both performance and memory wise when it comes to adding
properties. Most of the time, it evolves new type when properties are
added and it is less likely to get shared. Optimize so that if we are
deleting last property then transition to previous type of PathTypeHandler

Address issue #1713