[MERGE #1906 @Cellule] WASM - Imported wasm functions are used as is
Merge pull request #1906 from Cellule:wasm/import

A thunk is created for every imports. That thunk will do the conversion from primitive types to javascript types.
If the imported function is a WebAssembly function, the thunk is not used in that instance.
This also simplifies the internal model of the module because the function index now matches the representation the spec has.

Closes #1895
Closes #1887