[MERGE #1931 @leirocks] fix assertion on heapAllocation->isAllocationUsed = true
Merge pull request #1931 from leirocks:9634313

this is because while code gen client process crashed, and VM operation(virtual query) failed and exception thrown, and the state is not updated but the allocation is added to the list

then when RPC rundown kicks in then freeing the heap, it's in wrong state.

move the virtualQuery before adding to the list.
also make a fix for OOP JIT that we don't need to retry -- since we can't recover memory from the client process recycler