[MERGE #1944 @satheeshravi] Disabling StackArgs when the source is inside eval expression
Merge pull request #1944 from satheeshravi:StackArgsBugArgIns

This is a bug fix - where one of the assumptions was failing as an
assertion failure - This is true only when the source is inside an eval
expression.

Though the fix for this bug is really straight forward, I have decided to
disable the optimization when the source is inside an eval expression - to
minimize risk.