[MERGE #1930 @obastemur] Clang3.9: Fix name mangling issue
Merge pull request #1930 from obastemur:clang39s

Clang 3.9 is the default Clang on some linux package managers
and ChakraCore doesn't build with it.

IMHO, this is a Clang bug and I will file an issue separately.
However, this harmless (and esthetic :o) workaround
helps us not to trigger the issue.