[MERGE #2031 @digitalinfinity] Fix flakiness in serialized tests in ch
Merge pull request #2031 from digitalinfinity:ch_fix

Serialized bytecode tests started seeing flakiness after my ETW change.
This is because that change exposed a bug in ch- ch was freeing the serialized bytecode before disposing the runtime. We just happened to
not need any of the serialized data during our test runs before- with the ETW change, we started logging the name of the function body being collected and this would AV if the serialized bytecode data got decommited before
we called JsDisposeRuntime.

The simplest fix for now is to call JsDisposeRuntime before freeing the bytecode data