[MERGE #2110 @akroshg] Fixing on reentrant parsemodulesource.
Merge pull request #2110 from akroshg:mod2

When we get the ParseModuleSource while we fetch module - we go into a state where we don't notify the host about the error or ready state. That makes the module hung.
Fixed that by setting parent and child relation in the InitializeModuleRecord. That way when the ParseModuleSource re-entrancy happens and we can bubble up the error/ready state to parent properly.
Added native unittest for forcing the re-entrant parsemodulesource.