[MERGE #2117 @Penguinwizzard] Fix LdLen operand tags
Merge pull request #2117 from Penguinwizzard:two_lifetimes_fix

In the change to defer handling of length field loads until later in the optimizer, the need arose to move from a LdFld on a PropertySym SymOpnd to a StackSym RegOpnd (ldfld a->length to ldlen a). Unfortunately, in this transition a few flags were lost on the operand; the most notable is the isJITOptimized flag, which was causing an assert to be hit a good number of times.