Fix ProjectionTests after TypedArray ctor change
The change to the TypedArray constructors allowed any non-array like
object passed as the first to lead to creation of an empty TypedArray.

This test in the ProjectionTests expected this would fail which is no
longer the case.

Fix by removing the test.

https://github.com/microsoft/ChakraCore/pull/2016
ChakraCore commit b6161314131338cf0ae04bd8f88ff20e29132b39