[MERGE #2151 @ThomsonTan] Recheck array type before checking element is MissingItem after prototype lookup on the array
Merge pull request #2151 from ThomsonTan:FixMissingItemCheckOnArray

When slicing an array (Array.prototype.slice),  we make a shadow copy of original array, then tries to populate the MissingItems in the copied array with elements from the prototype chain of the source array. Lookup in the prototype chain could change its array type so we need recheck the array type after the lookup returns before retrieving remaining items as MissItems, or we get the wrong element to check against.