[MERGE #2168 @MikeHolman] relax the rpc failfast
Merge pull request #2168 from MikeHolman:relaxrpcfailfast

This failfast was a bit aggressive, and firing in scenarios where we don't necessarily want to get dumps (e.g. a user kills the server process as part of their shutdown). This excludes from firing a failfast for the known scenarios. Instead we will throw a codegen aborted error if this happened during codegen, or OOM in other cases.