[MERGE #2157 @MikeHolman] Use Sections for OOP JIT
Merge pull request #2157 from MikeHolman:sharemem

This change adds new SectionAllocWrapper classes, which manages Sections for JIT code pages.

When you call Alloc with these APIs, sections are mapped into the runtime process with PAGE_EXECUTE permissions. For SectionAllocWrapper, we maintain a map of outstanding sections modeled after the GC's HeapBlockMap for fast lookup.
If you want to map a section into the JIT process, you can call AllocLocal, which will map (a portion) of your section in as PAGE_READWRITE. It is the callers responsibility to call FreeLocal on the local section mapping.

Sections don't support decommit, which is used most importantly in PreReservedSectionAllocWrapper. Instead, to remove pages from the working set we ZeroMemory and set PAGE_NOACCESS protection on the runtime view.

InterpreterThunkEmitter was also modified to use sections. On the server we map everything in and then throw away the section handle. When the runtime is done, it can directly call to UnmapViewOfSection.
Now for OOP JIT we use the thunkBlocks lists for tracking outstanding interpreter thunks rather than maintaining an EmitBufferManager on the JIT process.