[MERGE #2216 @ThomsonTan] Don't remove dead syms which are referenced in callSequence in infinite loop blocks
Merge pull request #2216 from ThomsonTan:FixCallSequence

When function with infinite loop is inlined as parameter evaluation for another function, callSequence is propagated to infinite loop blocks in forward pass, but usually the operands in callSequence are not live because there is no way to exit the loop (so it is uncontional infinite loop). This triggers assertion when filling bailout from callSequence (because no value for given sym).

This change fixes it by not removing dead sym in symToValueMap if it is referenced in callSequence.