[MERGE #2307 @rajatd] Recover perf regression related to stack args
Merge pull request #2307 from rajatd:fixBox2dRegression

A perf regression had crept in with a refactor I did a while ago.

We dont have block-level merging of arguments object syms in the globopt. So, as much as possible, we try to keep the set of arguments object syms same across blocks. To this end, we ignore tracking the arguments object alias introduced by ByteCodeArgoutCapture. To do this, the globopt expects the argout instructions to be immediately preceding the InlineBuiltInStart instr for .apply.

The refactor I did had broken this assumption. Reverting to that.