[MERGE #2372 @leirocks] Broadly enable software write barrier
Merge pull request #2372 from leirocks:swb

This is to enable concurrent GC on non-Win32 platform.
On Windows we dependend on both software write barrier and hardware write barrier(aka WriteWatch)  to do concurrent GC.
On Linux the write watch is not supported.
This change is to make software write barrier globaly enabled for recycler allocated stcutures  and hence enable concurrent GC on Linux.