[MERGE #2396 @boingoing] Some string APIs operating on UTF-8 buffers do not property compute the length of the UTF-8 buffer in bytes
Merge pull request #2396 from boingoing:UnifiedRegexDecodeUnicodeSequences

Many of our string APIs which take LPCUTF8 buffers do not know how many
bytes are in those buffers. Typically, these APIs need to walk over the
characters in the decoded string so they take in parameters for the
number of decoded characters. But the number of decoded characters is
not necessarily the same as the number of bytes. This is the case when
there are Unicode sequences encoded into the UTF-8 string.

As a consequence of this missing info, APIs such as utf8::CharsAreEqual
made the assumption that the UTF-8 buffer always contains at least 4
bytes - the maximum number of bytes in a Unicode sequence - without
knowing if that was actually true or not. It turned out that this was
not always true.

This change computes the correct byte length for the APIs which
previously made assumptions about the number of bytes in the buffer.

Fixes:
https://microsoft.visualstudio.com/OS/_workitems?id=10440534&_a=edit