[MERGE #2409 @pleath] Let redeferral work around block scope limitation of deferred parsing.
Merge pull request #2409 from pleath:owa

Detect cases where only block-scoping prevents upfront deferral. Also detect cases where it is safe to treat a lexically-scoped function as a var-scoped one. For functions that meet both criteria, use the var binding for it in place of the lexical binding. If at byte code gen time there are no lexical/with/catch scopes in the current chain, mark the function as a redeferral candidate.