[MERGE #2477 @curtisman] Fix lowering of BrOnNotEmpty
Merge pull request #2477 from curtisman:fixlower

When we finish iterating the ForInCache in JIT'ed code and if the cache is completed, we can exit the loop.
For BrOnNotEmpty, the target of the branch is the loop body instead of loop end.  We need to use the label after the instruction