[PR #2501 @LouisLaf] Fix kraken regressions [Update submodule only]
My previous checking caused a regressions in a couple of the Kraken crypto benchmarks. Loading int32 from native arrays is ok even if int-type-spec is disabled, because we will convert these to float64, as long as float-type-spec is enabled.

[Core Commit Message]

[MERGE #2501 @LouisLaf] Fix kraken regressions

Merge pull request #2501 from LouisLaf:kraken

My previous checking caused a regressions in a couple of the Kraken crypto benchmarks. Loading int32 from native arrays is ok even if int-type-spec is disabled, because we will convert these to float64, as long as float-type-spec is enabled.