[MERGE #2267 @ThomsonTan] Fix inconsistent type spec for Math.abs(INT32_MIN) in loops
Merge pull request #2267 from ThomsonTan:FixMathAbs

When there is Math.abs(INT32_MIN) in loop, we do int type spec on it in LoopPrePass, and const fold it after that. But const fold changes its dst type to float because Math.abs(INT32_MIN) is not valid int32 value. This inconsistency causes error if dst of Math.abs(INT32_MIN) flows back loop header then lossy conversion is needed on the float var. The lossy conversion violates the original assumption in loop header that the var is lossless.