[MERGE #2516 @akroshg] Proxy fixes when it is revoked while in trap
Merge pull request #2516 from akroshg:proxy1

When we are proxy trap, it is possible that we invoke the user script and revoke current proxy.
In that case we made the target and handler to be nullptr. After revoking
once we return back we probably be working on null object.
According to spec we need to take the local variables when we begin in the
trap, that way the target and handler will be not be gone.
Added test cases for all traps.