[MERGE #2534 @pleath] Repair incorrect fix for double increment of literal regex count.
Merge pull request #2534 from pleath:10368633

On redeferral of an enclosing function, where nested functions are still in a compiled state, we do not want to re-increment the literal regex count for the nested functions. But we do still want to count the regexes and allocate new regex storage for the function we're recompiling. A new simple fix handles both cases.