[MERGE #2531 @satheeshravi] Fixing !ldsym in jd for OOPjit
Merge pull request #2531 from satheeshravi:release1.4_ChakraDiag

We don't register the native addresses and interpreter thunk addresses for OOPjit scenarios - since the allocations linked list is not populated by the runtime process.

Fix:
Used the cache that we now populate on the script context to register the address.
For interpreter thunk emitter - we use the thunkBlocks linked list.