[MERGE #2596 @akroshg] ResetObject is setting undefined for internal properties
Merge pull request #2596 from akroshg:reset_1.4

Once the GetInternalProperty fails the stackTraceValue (and others) are set to GetMissingPropertyResult. Later below in that function we  set the undefined object for such property which actually leads to crashes in the Weakmap functionality as we never expect it to be undefined.
Fixed this and put an assert just incase in the Weakmap we get undefined again.