[MERGE #2632 @ThomsonTan] Fix Hoisting ValueType of String concat
Merge pull request #2632 from ThomsonTan:HoistConcat1

When hoisting `SetConcatStrMultiItemBE` to loop landingpad, the type of src operands could be loosed to be non-string (this was ensured by some bailout in the loop, but the type guarantee will disappear in landing pad if the bailout was not also hoisted). In this case we need add `Conv_PrimStr` back to make sure `String` is passed to `SetConcatStrMultiItemBE` as src operands.