[MERGE #2646 @akroshg] Module export field issue with namespace.
Merge pull request #2646 from akroshg:module2

The export function is not put to the module namespace object as the logic of namespace's initialize has an issue. We should be putting the exports to the namespace if we don't find it in the importlist. However the logic was saying otherwise. Fixed that.