[MERGE #2657 @suwc] Clean up JavascriptArray::DirectGetItemAt and ::DirectSetItemAt
Merge pull request #2657 from suwc:build/suwc/bugfix

BigIndex overloads of JavascriptArray::DirectGetItemAt and ::DirectSetItemAt are not helpful for the goal of readable and secure codepaths.
Remove all calls to the BigIndex overload of DirectGetItemAt.
Create an overloaded wrapper for DirectSetItemAt named GenericDirectSetItemAt.
Replace the handful of calls to BigIndex flavor of DirectSetItemAt to calls to the new wrapper.