[MERGE #2654 @Cellule] WASM: Use wabt in chakra
Merge pull request #2654 from Cellule:wasm/wabt

We've been lacking a good way to convert wast to wasm. In the past we've used wabt to convert wast files and checkin the compiled version.
Since Wabt looks like something to stay and is reliable, we decided to integrate it directly in chakra to help do testing.
Right now the only api exposed is `WebAssembly.wabt.convertWast2Wasm(<text>)` but in the near future I would like to expose a method to convert spec like wast files with asserts.