[MERGE #2698 @pleath] Fix bad code with call to closure-captured local function from inside 'with'.
Merge pull request #2698 from pleath:11295796

The case where the with statement is in a function nested within the function that instantiates the called function is already handled. Now that functions nested within with can be deferred, we need to handle the case where the with is in the enclosing function itself.