[MERGE #2703 @kunalspathak] Reduce memory consumption of Utf8SourceInfo objects
Merge pull request #2703 from kunalspathak:scriptSourceFix

Today we hold on to `scriptSource` objects inside `Utf8SourceInfo` even if we make a copy of the script source inside source holder. This is unnecessary and increase the working set. The fix is to not hold reference anymore.

However if the `scriptSource` is external array buffer and is managing its own life time, we don't need to make copy of it and instead we just hold the reference to `scriptSource` for scenarios like `function.name.toString()` to work (which is how it is done today).