[MERGE #2687 @Cellule] WASM: Validate unreachable code
Merge pull request #2687 from Cellule:users/micfer/wasm/unreachable

According to spec we are supposed to validate unreachable code as well.

I decided to do the work for an noop byte code writer as part of this change because it was too hard and pointless to try to emit valid bytecode in unreachable cases. Therefore, when we switch to unreachable mode, we change the bytecode writer so subsequent calls to write will do nothing.
In a future PR I will make `WebAssembly.validate` solely use the noop byte code writer to speed up validation #2688.