[MERGE #2629 @MikeHolman] fold pattern of a+(b<<pow2) into lea
Merge pull request #2629 from MikeHolman:shiftaddtolea

This pattern shows up a fair bit in wasm and asm.js code. It's about a 3% win in zlib, 2% in jetstream, and 1% in wasm unity.

I added peep to our PreLowerPeeps phase for catching this.

It also required adding support in the encoder and couple other places to allow IndirOpnds to not have a baseOpnd post-lower.