[MERGE #2710 @pleath] Fix bad byte code when block-scoped function is called from with.
Merge pull request #2710 from pleath:11291207

Yesterday's fix was incomplete, because it failed to address the case in which the nested function called from a with statement is declared in a lexical scope. The concept we're conservatively using, of a scope containing a with statement as well as closure-captured data, has to be tracked on individual scopes rather than functions.