[MERGE #2613 @leirocks] fix assertion on AuxPtr
Merge pull request #2613 from leirocks:9779807

While fixing a dead lock bug and adding assertion for that the work item queue should not be locked at the time of getting the auxPtrs. The assertion is wrong when JIT is collecting data in foreground thread causes stack probing and hence call dispose then re-enter script, which can call GetAuxPtrsWithLock. Fixing with checking if JIT is in foreground.
Also change to not call dispose in codeGen/gathering data, to prevent unpredicted reentrant issue.

Fix the real possible points of dead lock on etw rundown and remove IsLockedByAnyThread (which is hacky and not accurate, should not allow the etw rundown thread to updated the counters on function body, if any we should fix it)