[PR #2726 @dilijev] Migrate Intl tests to ChakraCore.
Migrate Intl tests to ChakraCore.

* Migrated all of ProjectionTests/Tests/Intl/* to Core
* Deduplicated these tests with the unittest/Intl directory

Most tests which were moved to core were re-written to work with
UnitTestFramework and to remove the dependency on the ProjectionTest framework
(most notably in the form of projectionsglue.js which could be replaced by
UnitTestFramework in a straightforward way).

Some tests could not be easily migrated and there was not much benefit of
initially doing so. Tests which should eventually be moved to ChakraCore now
reside in IntlCore.

The one remaining test in unittest/Intl is ValidateGeneratedByteCode.cmd and is
disabled. Removing or re-enabling this is out of scope of the current change
but will come up in a later unit test audit we're planning.

[Core Commit Message]

[MERGE #2726 @dilijev] Migrate Intl tests to ChakraCore.

Merge pull request #2726 from dilijev:intltest

Closes #2726