[MERGE #2821 @pleath] Fix cases of bad binding in arrow function parameter lists.
Merge pull request #2821 from pleath:11557877

Fix some cases where block enumeration differed between buildAST and !buildAST compiles. If the initial parse of an arrow function parameter was deferred, and the reparse on discovery of the arrow token was (necessarily) not deferred, then the inconsistent state of the pid refs left by the initial parse led to incorrect bindings.