Fixing ARM64 projection
When the calling convention runs out of registers it will start
pushing parameters onto the stack. In this case we were moving the
stack pointer to make room for the parameters, but never restoring
the stack pointer before returning. This caused the first two
parameters to be restored to lr/fp which crashes on return.